requirements.txt
streamlit
langchain
openai
mysql-connector-python
reportlab
